+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Addition of a precompile step to woodpecker PR check
+### Fixed
+- fix type error due to bad tsconfig
 ### Changed
 - Woodpecker: do not run changelog-check when PR contains `dependabot` label
 - Made the number variable also show placeholders
@@ -21,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bumps `eslint-plugin-ember` from 11.4.6 to 11.9.0
 - Bumps `@typescript-eslint/eslint-plugin` from 5.45.0 to 5.60.1
 - Bumps `ember-velcro` to 2.1.0
+- Bumps `@lblod/ember-rdfa-editor` to 4.0.0
 
 ## [8.2.2] - 2023-06-28